增加修改课程设备接口

+++ src/main/java/com/remoteLaboratory/service/Impl/CourseDeviceServiceImpl.java
@@ -107,6 +107,9 @@ public class CourseDeviceServiceImpl implements CourseDeviceService {
                 if(!loginUser.getUserType().equals(Constants.USER_TYPE_ADMIN) && !course.getTeacherId().equals(loginUser.getId())) {
                     throw new BusinessException(Messages.CODE_50200);
                 }
+                if(course.getStatus().equals(1)) {
+                    throw new BusinessException(Messages.CODE_40010, "课程正在进行不能删除！");
+                }
                 courseDeviceRepository.delete(id);
                 LogRecord logRecord = new LogRecord();
                 logRecord.setType("删除");
